pyproject.toml
[project]
name = "fetching-virustotal-file-report"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "typer>=0.15.3",
    "ujson>=5.10.0",
]
